+++ src/RestrictedBoltzmannMachines.jl
@@ -7,8 +7,9 @@ using EllipsisNotation: (..)
 using FillArrays: Falses, Zeros, Ones, Trues
 using LinearAlgebra: Diagonal, dot, logdet
 using LogExpFunctions: log1pexp, logaddexp, logistic, logit, logsumexp, softmax
+using MLUtils: DataLoader
 using Optimisers: AbstractRule, Adam, setup, update!
-using Random: AbstractRNG, default_rng, rand!, randexp, randn!, randperm
+using Random: AbstractRNG, default_rng, rand!, randexp, randn!
 using SpecialFunctions: erf, erfcx, logerfcx
 using Statistics: mean